fix: move metadata to layout
