+++ src/app/[lang]/about/become-a-collaborator/page.tsx
@@ -1,53 +1,3 @@
-import {
-  getDictionary,
-  type Locale,
-} from "@/internationalization/dictionaries";
-import { type Metadata } from "next";
-
-export function generateMetadata({
-  params: { lang },
-}: {
-  params: { lang: Locale };
-}): Metadata {
-  const dict = getDictionary(lang);
-
-  return {
-    title: dict.seo.become_a_collaborator.title,
-    description: dict.seo.become_a_collaborator.description,
-    keywords: [
-      "student center",
-      "engeneering",
-      "informatics",
-      "uminho",
-      "university",
-      "students",
-      "CeSIUM",
-      "CeSIUM UMinho",
-    ],
-    openGraph: {
-      url: "https://cesium.di.uminho.pt/about/become-a-collaborator",
-      type: "website",
-      title: dict.seo.become_a_collaborator.title,
-      description: dict.seo.become_a_collaborator.description,
-      images: [
-        {
-          url: "https://cesium.di.uminho.pt/og.png",
-          width: 1200,
-          height: 630,
-          alt: "cesium.di.uminho.pt",
-        },
-      ],
-    },
-    alternates: {
-      canonical: "https://cesium.di.uminho.pt/about/become-a-collaborator",
-      languages: {
-        en: "https://cesium.di.uminho.pt/en_US/about/become-a-collaborator",
-        pt: "https://cesium.di.uminho.pt/pt_PT/about/team",
-      },
-    },
-  };
-}
-
 export default function BecomeACollaborator() {
   return <main>Become a Collaborator</main>;
 }
