+++ src/app/[lang]/about/departments/page.tsx
@@ -9,56 +9,6 @@ import { CardType, type TeamData } from "@/lib/types";
 import { fetchTeamData } from "@/lib/utils";
 import { useState, useEffect } from "react";
 
-import {
-  getDictionary,
-  type Locale,
-} from "@/internationalization/dictionaries";
-import { type Metadata } from "next";
-
-export function generateMetadata({
-  params: { lang },
-}: {
-  params: { lang: Locale };
-}): Metadata {
-  const dict = getDictionary(lang);
-
-  return {
-    title: dict.seo.departments.title,
-    description: dict.seo.departments.description,
-    keywords: [
-      "student center",
-      "engeneering",
-      "informatics",
-      "uminho",
-      "university",
-      "students",
-      "CeSIUM",
-      "CeSIUM UMinho",
-    ],
-    openGraph: {
-      url: "https://cesium.di.uminho.pt/about/departments",
-      type: "website",
-      title: dict.seo.departments.title,
-      description: dict.seo.departments.description,
-      images: [
-        {
-          url: "https://cesium.di.uminho.pt/og.png",
-          width: 1200,
-          height: 630,
-          alt: "cesium.di.uminho.pt",
-        },
-      ],
-    },
-    alternates: {
-      canonical: "https://cesium.di.uminho.pt/about/departments",
-      languages: {
-        en: "https://cesium.di.uminho.pt/en_US/about/departments",
-        pt: "https://cesium.di.uminho.pt/pt_PT/about/departments",
-      },
-    },
-  };
-}
-
 export default function Departments() {
   const dict = useDictionary();
