+++ src/app/[lang]/events/page.tsx
@@ -1,54 +1,3 @@
-import {
-  getDictionary,
-  type Locale,
-} from "@/internationalization/dictionaries";
-import { type Metadata } from "next";
-
-export function generateMetadata({
-  params: { lang },
-}: {
-  params: { lang: Locale };
-}): Metadata {
-  const dict = getDictionary(lang);
-
-  return {
-    title: dict.seo.events.title,
-    description: dict.seo.events.description,
-    keywords: [
-      "student center",
-      "engeneering",
-      "informatics",
-      "uminho",
-      "university",
-      "students",
-      "CeSIUM",
-      "CeSIUM UMinho",
-      "events",
-    ],
-    openGraph: {
-      url: "https://cesium.di.uminho.pt/events",
-      type: "website",
-      title: dict.seo.events.title,
-      description: dict.seo.events.description,
-      images: [
-        {
-          url: "https://cesium.di.uminho.pt/og.png",
-          width: 1200,
-          height: 630,
-          alt: "cesium.di.uminho.pt",
-        },
-      ],
-    },
-    alternates: {
-      canonical: "https://cesium.di.uminho.pt/events",
-      languages: {
-        en: "https://cesium.di.uminho.pt/en_US/events",
-        pt: "https://cesium.di.uminho.pt/pt_PT/events",
-      },
-    },
-  };
-}
-
 export default function Events() {
   return <main>Events</main>;
 }
